fix lua call bp function may cause crash  fix loadclass function for construct  refactor test case code
